refactor: declutter source files, comments and tests
